# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Todo Project CRUD",
    "description": "My Todo Project Application API"
  },
  "host": "localhost:8000",
  "basePath": "/",
  "tags": [
    {
      "name": "Todo",
      "description": "API for todos"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/todos": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint to get the todos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint to add a todo",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Todo information.",
            "schema": {
              "$ref": "#/definitions/AddTodo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint to get the todo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint to update a todo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "Todo information.",
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Todo"
        ],
        "description": "Endpoint to delete the todo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "id"
        },
        "definition": {
          "type": "string",
          "example": "definition"
        },
        "comment": {
          "type": "string",
          "example": "commen"
        },
        "date": {
          "type": "string",
          "example": "2022-08-25T12:25:36.874Z"
        },
        "done": {
          "type": "boolean",
          "example": true
        }
      },
      "required": [
        "id",
        "definition"
      ]
    },
    "Todos": {
      "$ref": "#/definitions/Todo"
    },
    "AddTodo": {
      "type": "object",
      "properties": {
        "done": {
          "type": "boolean",
          "example": true
        },
        "definition": {
          "type": "string",
          "example": "definition"
        },
        "comment": {
          "type": "string",
          "example": "commen"
        },
        "date": {
          "type": "string",
          "example": "2022-08-25T12:25:36.874Z"
        }
      },
      "required": [
        "definition"
      ]
    }
  }
}